实时数据结构修改，PreconditionResp修改

+++ src/main/java/com/hp/data/bean/tbox/DataResendRealTimeMes.java
@@ -18,7 +18,7 @@ public class DataResendRealTimeMes extends TBoxUpBean{
     private Short leftRearTirePressure;
     private Short rightFrontTirePressure;
     private Short rightRearTirePressure;
-    private Short windowInformation;
+    private Integer windowInformation;
     private Short vehicleTemperature;
     private Short vehicleOuterTemperature;
     private Short doorInformation;
@@ -155,11 +155,11 @@ public class DataResendRealTimeMes extends TBoxUpBean{
         this.rightRearTirePressure = rightRearTirePressure;
     }
 
-    public Short getWindowInformation() {
+    public Integer getWindowInformation() {
         return windowInformation;
     }
 
-    public void setWindowInformation(Short windowInformation) {
+    public void setWindowInformation(Integer windowInformation) {
         this.windowInformation = windowInformation;
     }
